resolves #128 allow to extend the default document converter
