+++ lib/cli.js
@@ -60,7 +60,7 @@ class PdfInvoker extends Invoker {
     if (templateRequireLib) {
       templates = Invoker.requireLibrary(templateRequireLib)
     } else {
-      templates = require('./document/templates.js')
+      templates = require('./document/document-converter').templates
     }
     converter.registerTemplateConverter(processor, templates)
     Invoker.prepareProcessor(args, processor)
