+++ test/templates_test.js
@@ -8,7 +8,7 @@ chai.use(dirtyChai)
 
 const asciidoctor = require('@asciidoctor/core')()
 const converter = require('../lib/converter.js')
-const templates = require('../lib/document/templates.js')
+const { templates } = require('../lib/document/document-converter')
 converter.registerTemplateConverter(asciidoctor, templates)
 
 describe('Default converter', () => {
